+++ jaclang/core/llms/ollama.py
@@ -1,0 +1,92 @@
+"""Ollama client for MTLLM."""
+import re
+
+PROMPT_TEMPLATE = """
+[System Prompt]
+This is an operation you must perform and return the output values. Neither, the methodology, extra sentences nor the code are not needed.
+Input/Type formatting: Explanation of the Input (variable_name) (type) = value
+
+[Type Explanations]
+{type_explanations}
+
+[Information]
+{information}
+
+[Inputs Information]
+{inputs_information}
+
+[Output Information]
+{output_information}
+
+[Action]
+{action}
+
+{reason_suffix}
+"""  # noqa E501
+
+WITH_REASON_SUFFIX = """
+Reason and return the output results(s) only such that <Output> should be eval(<Output>) Compatible and reflects the expected output type, Follow the format below to provide the reasoning for the output result(s).
+
+[Reasoning] <Reasoning>
+[Output] <Output>
+"""
+
+WITHOUT_REASON_SUFFIX = """Return the output result(s) only such that <Output> should be eval(<Output>) Compatible and reflects the expected output type, Follow the format below to provide the output result(s).
+
+[Output] <Output>
+"""  # noqa E501
+
+
+class Ollama:
+    """Ollama API client for Large Language Models (LLMs)."""
+
+    MTLLM_PROMPT: str = PROMPT_TEMPLATE
+    MTLLM_REASON_SUFFIX: str = WITH_REASON_SUFFIX
+    MTLLM_WO_REASON_SUFFIX: str = WITHOUT_REASON_SUFFIX
+
+    def __init__(self, **kwargs: dict) -> None:
+        """Initialize the Ollama API client."""
+        import ollama  # type: ignore
+
+        self.client = ollama.Client(host=kwargs.get("host", "http://localhost:11434"))
+        self.model_name = kwargs.get("model_name", "phi3")
+        self.default_model_params = {
+            k: v for k, v in kwargs.items() if k not in ["model_name", "host"]
+        }
+
+    def __infer__(self, meaning_in: str, **kwargs: dict) -> str:
+        """Infer a response from the input meaning."""
+        model = str(kwargs.get("model_name", self.model_name))
+        if not self.check_model(model):
+            self.download_model(model)
+        model_params = {k: v for k, v in kwargs.items() if k not in ["model_name"]}
+        messages = [{"role": "user", "content": meaning_in}]
+        output = self.client.chat(
+            model=model,
+            messages=messages,
+            options={**self.default_model_params, **model_params},
+        )
+        return output["message"]["content"]
+
+    def check_model(self, model_name: str) -> bool:
+        """Check if the model is available."""
+        try:
+            self.client.show(model_name)
+            return True
+        except Exception:
+            return False
+
+    def download_model(self, model_name: str) -> None:
+        """Download the model."""
+        self.client.pull(model_name)
+
+    def resolve_output(self, meaning_out: str, did_reason:bool) -> dict:
+        """Resolve the output string to return the reasoning and output."""
+        print(meaning_out)
+        reasoning_match = re.search(r"\[Reasoning\](.*)\[Output\]", meaning_out)if did_reason else None
+        output_match = re.search(r"\[Output\](.*)", meaning_out)
+        reasoning_match = reasoning_match.group(1).strip() if reasoning_match else None
+        output_match = output_match.group(1).strip() if output_match else None
+        if not output_match and not reasoning_match:
+            output_match = meaning_out.strip()
+        return {"reasoning": reasoning_match, "output": output_match}
